Refactor validators
Created a new file to separate validators and adjust their references.
